requirements-web.txt
# AIO 潛力分析器 Web 版本依賴套件
# ==========================================

# Django 框架與相關套件
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-oauth-toolkit>=1.7.0
django-allauth>=0.57.0

# 數據庫
psycopg2-binary>=2.9.7
django-extensions>=3.2.3

# 任務隊列
celery>=5.3.0
redis>=4.6.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# Google APIs (原有)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
google-ads>=22.1.0

# 數據處理 (原有)
pandas>=2.0.0
numpy>=1.24.0

# 異步處理 (原有)
aiohttp>=3.8.0
asyncio-throttle>=1.0.2

# 進度條和用戶界面 (原有)
tqdm>=4.65.0

# 配置和環境管理 (原有)
python-dotenv>=1.0.0
pyyaml>=6.0

# 日誌記錄 (原有)
colorlog>=6.7.0

# Web 相關
whitenoise>=6.6.0
gunicorn>=21.2.0
channels>=4.0.0
channels-redis>=4.1.0

# 安全性
cryptography>=41.0.0
django-ratelimit>=4.1.0

# 監控與調試
django-debug-toolbar>=4.2.0
sentry-sdk>=1.32.0

# 文件處理
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# 測試工具 (開發用)
pytest>=7.4.0
pytest-django>=4.5.0
pytest-asyncio>=0.21.0
pytest-celery>=0.0.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# 代碼品質工具 (開發用)
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.4.0

# API 文檔
drf-spectacular>=0.26.0

# 快取
django-redis>=5.4.0
